Don't create database in serving command
Before, we were creating the database in the
`livetestserver` command. This worked, but it
meant that the `test` command had to wait
to start until `livetestserver` was done migrating
the database. If we instead move the database
setup to the `test` command, then it can
start whenever the database is setup.